Remove openbsd pf state flush cmds
The command `pfctl -a anchor -F states`, as of OpenBSD v6.8, flushes all
existing states, regardless of the specification of an anchor, presently
rendering said command unfit for this program.

The purpose of the command was to terminate ongoing outbound traffic to
newly blacklisted IP addresses upon the creation of a new ruleset.